Added Thunder
Added Thunder at random intervals, a random number between 0 & 555, each time that number is exactly 1, a thunder is generated.

This only works when the rain is blue.

Also simplified the way color is calculated in some cases.
Added some variables for actual thunder bolts but removed them and let them be comments.